Now quantizes the image for display at least
Using quantizr

I guess from here it's time to start working on OSC things?
Although perhaps we should take a crack at 4-bit grayscale first. I'm
not sure if quantizr can handle that though...

Also I realize that the 8-bit floats in VRC might represent a
precision issue when we want to be sending u8s. We might end up having
to use integers. Downside is then we have to make an animator to
decode all that shit....