Fix for problems with threaded title setting and alert dialog boxes: Use channels
This is better than the hacky thing I did before. We do it for alert
dialog boxes.

As a bonus we even make a custom panic handler hook that makes a
dialog box. Maybe we could actually use panic as an OK thing to do in
the bg thread to indicate issue with loading the image or something?
I'm not sure I like it though.