We need to wake fltk when sending messages
The built-in channel in fltk-rs already does this. But I moved away
from it for some reason (because I want to make the processing thread
a continously running bg thread and that should also have a channel,
so might as well use the same everywhere I thought. OTOH we could have
the fltk-rs one to send things to the app, and the other one to send
things to the BG thread I guess?