Refactor so that we use a background process that listens on a channel for messages
The idea is this way we do not need a global imagepath etc. and we
avoid spawning tons of threads.

The problem: Because of the type of channel we use a lot of
UpdateImage messages gets queued up unneccesarily when pulling a
slider, keeping on updating the image. However we don't see it go
slowly through all the configuration changes as you would have
expected as we are relying on global state/directly accessing the
various widgets. (Maybe we should have UpdateImage take a struct
though? Although it would mean even worse behaviour but less
surprising from a programming stand-point)

The solution would likely be to implement some sort of interrupting of
the current task whenever a new configuration appears. Is this
possible somehow using async? With a nice-to-have on top of that being
that we pass the entire config through the channel instead of using
shared memory.

Of course one thing to try might just be using a
std::mpsc::sync_channel with a size of one or so. But that will likely
have the not-so-nice behaviour of freezing the entire program while we
are doing the heavy processing and all that. That being the entire
reason to use threading to begin with.