Implement a .recv method for the message queue that gives us only the one
This works under the assumption that there is only ever a single
consumer. Because the .recv method would steal from other potential
consumers, even if it only wants one thing out of the queue.

Now this also makes it so MessageQueueReceiver has to be mut to be
able to be used.

Maybe I should've implemented some iterator version of this instead.
Maybe I could implement some stupid type of iterator on top of the
.recv method. Who knows.