Use receiver.recv() from the MessageQueue instead of drain
Thinking about it, this should be able to be more responsive in
certain cases. Consider that the background process might gobble up
something + an UpdateImage. Then while processing the "something" an
UpdateImage gets added to the queue. Which means now two UpdateImage
messages are basically set to run after one another.

If the background process instead only takes one at a time, there is a
higher chance that an UpdateImage will get overwritten with a newer
UpdateImage while it is processing the "something" (say a LoadImage)