Simplify code to make the Send OSC window
Now that the window deletation is initiated from (but not run on) the
Send OSC BG-thread, it is back to just storing to cancel_flag, so we
don't need cancel_cb any more really.

Perhaps it'd be even more simplifying to just pre-create this window
and show/hide it when needed instead though?