add the png crate
For now let's just use the PNG crate directly to save our indexed
images.

Do consider trying out the ril crate as a potential replacement for
the image crate though. That seems to be a generic imaging library,
but with better support for paletted images. It also has its own
quantiser as well (including lossy ones and stuff, could be faster and
    all that!)