save_png: Fix the issues when width doesn't divide cleanly for bit depth
For 1bpp it only worked for images with a width that is divisible by 8
For 2bpp it was divisible by 4
for 4bpp it was divisible by 2
For 8bpp it was divisible by 1 (just works)

Fixed by looping per line. The accesses using .get fromn earlier fixes
the padding needed alright. Perhaps not the most efficient
implementation here, but it's easy