Adding support for lower bitdepths in send_osc (still grayscale only)
Copying code from save_png here.
I'd like to de-duplicate this later. Maybe possible to do
intelligently using Cow, to avoid cloning wheh not neccessary in
save_png?
Or do we just make it return a Vec<u8> always, but sometimes it's the
one that was passed in or so? Difficult lifetime-wise it is.