send_osc: Rewrite palette sending again with less hard-coding and reference usage
Also now we use a loop. I think this is pretty close to how I would've
written this in regular old C.

I should keep it in mind to KISS sometimes like this (although the
original .flat_map-based implementation was cool, I'd need something
that can collect an iterator into an array without heap).

This commit is basically training for future work in a no_std environment