Make default window size a little bit taller
Needed to accomodate all the buttons I've added.

We really need a scrollbar type thing I guess now in case the window
becomes too small (well. there is a lot of things that could be
handled better when too small. Perhaps the best solution is to pick
something else than fltk, but that is a lot of effort rewriting code)