Add an option whether to resize to fill or resize to fit
Currently results in the height or width changing though, which is not
ideal for the OSC sending use case. I will add padding in a later
commit.

Probably best to add the padding after quantization so we do not end
up always requiring black to be part of the palette.

For the ResizeType enum I use strum to be able to convert it to/from
string. This is a first to try strum out. Although maybe I am better
of writing my own macros eventually. The enums in send_osc.rs cannot
really be handled too well with only strum, and require some custom
code.