main: Change sizes of some buttons and things if the screen is small
We pack things a bit more aggressively now on a smaller screen, while
allowing it all to be a bit roomier on a roomier screen.

This also standardized the sizes of the various widgets I used. Before
it was a bit inconsistent with some toggles being 30 pixels tall and some being
20 pixels tall.