Add heuristic to choose padding color
Essentially we just look at all pixels on the sides and use the color
there is the most of for a reasonable heuristic to match the
background color in the image itself, which can hopefully improve RLE
compression down the line.

At least for the tall case, the heuristic could become even smarter by
for each row actually checking all pixels from the edge going inward
until the first pixel change and throwing those into the count as
well. But that would mean checking more pixels and more complexity for
little benefit.

The heuristic matters very ilttle for the wide case, and not at all
for the square case as no padding will be added.